Replace SlintWgpuCanvas with more general one
