+++ map/src/lib.rs
@@ -1,9 +1,9 @@
 extern crate core;
 
 use crate::camera::{Camera, CameraController};
-use crate::route::RouteCosting;
 use crate::kml_viewer_group::KmlGroup;
 use crate::puck_group::SimplePuck;
+use crate::route::RouteCosting;
 use crate::tiles::tile_data::TileData;
 use crate::tiles::tiles_provider::{TilesMessage, TilesProvider};
 use futures::executor::block_on;
@@ -11,24 +11,24 @@ use futures::{pin_mut, Stream, StreamExt};
 use geo_types::private_utils::get_bounding_rect;
 use geo_types::{coord, Coord, Point, Rect};
 use geo_types::{LineString, Polygon};
+use glam::{DVec2, DVec3, Vec2};
+use num::clamp;
+use osm::styles::style_loader::StyleLoader;
+use osm::styles::{DashStyle, RenderStyle};
 use renderer::canvas_api::CanvasApi;
+use renderer::mesh_layers::feature_layers::FeatureLayerTag;
 use renderer::modifier::render_modifier::SpatialData;
 use renderer::render_group::RenderGroup;
 use renderer::renderer_api::RendererApi;
+use renderer::styles::style_id::StyleId;
 use renderer::{Renderer, RendererUpdateData, ShashlikRenderer};
 use route::route_controller::RouteController;
 use std::mem;
 use std::path::PathBuf;
 use std::sync::{Arc, LazyLock};
 use std::thread::spawn;
-use glam::{DMat4, DVec2, DVec3, Vec2};
-use num::clamp;
-use osm::styles::{DashStyle, RenderStyle};
-use osm::styles::style_loader::StyleLoader;
 use ttf_parser::Face;
 use wgpu::Texture;
-use renderer::mesh_layers::feature_layers::FeatureLayerTag;
-use renderer::styles::style_id::StyleId;
 use wgpu_canvas::wgpu_canvas::WgpuCanvas;
 
 mod camera;
